--- a/src/bitmap.cpp
+++ b/src/bitmap.cpp
@@ -212,29 +212,51 @@ bool Bitmap::GetTransparent() const {
 	return format.alpha_type != PF::NoAlpha;
 }
 
-ImageOpacity Bitmap::ComputeImageOpacity(Rect rect) const {
-	bool all = true;
-	bool any = false;
+ImageOpacity Bitmap::ComputeImageOpacity() const {
+	bool all_opaque = true;
+	bool all_transp = true;
 
-	DynamicFormat format(32,8,24,8,16,8,8,8,0,PF::Alpha);
-	std::vector<uint32_t> pixels;
-	pixels.resize(rect.width * rect.height);
-	Bitmap bmp(reinterpret_cast<void*>(&pixels.front()), rect.width, rect.height, rect.width*4, format);
-	bmp.Blit(0, 0, *this, rect, Opacity::Opaque());
+	auto* p = reinterpret_cast<const uint32_t*>(pixels());
+	const auto mask = pixel_format.rgba_to_uint32_t(0, 0, 0, 0xFF);
 
-	for (std::vector<uint32_t>::const_iterator p = pixels.begin(); p != pixels.end(); ++p) {
-		if ((*p & 0xFF) != 0)
-			any = true;
-		else
-			all = false;
-		if (any && !all)
-			break;
+	int n = GetSize() / sizeof(uint32_t);
+	for (int i = 0; i < n; ++i ) {
+		auto px = p[i] & mask;
+		all_opaque &= (px == mask);
+		all_transp &= (px == 0);
 	}
 
 	return
-		all ? ImageOpacity::Opaque :
-		any ? ImageOpacity::Partial :
-		ImageOpacity::Transparent;
+		all_transp ? ImageOpacity::Transparent :
+		all_opaque ? ImageOpacity::Opaque :
+		ImageOpacity::Partial;
+}
+
+ImageOpacity Bitmap::ComputeImageOpacity(Rect rect) const {
+	bool all_opaque = true;
+	bool all_transp = true;
+
+	const auto full_rect = GetRect();
+	rect = full_rect.GetSubRect(rect);
+
+	auto* p = reinterpret_cast<const uint32_t*>(pixels());
+	const int stride = pitch() / sizeof(uint32_t);
+	const auto mask = pixel_format.rgba_to_uint32_t(0, 0, 0, 0xFF);
+
+	int xend = (rect.x + rect.width);
+	int yend = (rect.y + rect.height);
+	for (int y = rect.y * stride; y < yend * stride; y += stride) {
+		for (int x = rect.x; x < xend; ++x) {
+			auto px = p[x + y] & mask;
+			all_transp &= (px == 0);
+			all_opaque &= (px == mask);
+		}
+	}
+
+	return
+		all_transp ? ImageOpacity::Transparent :
+		all_opaque ? ImageOpacity::Opaque :
+		ImageOpacity::Partial;
 }
 
 void Bitmap::CheckPixels(uint32_t flags) {
@@ -265,7 +287,7 @@ void Bitmap::CheckPixels(uint32_t flags) {
 	if (flags & Flag_ReadOnly) {
 		read_only = true;
 
-		image_opacity = ComputeImageOpacity(GetRect());
+		image_opacity = ComputeImageOpacity();
 	}
 }