--- a/src/bitmap.h
+++ b/src/bitmap.h
@@ -151,14 +151,6 @@ public:
 		Flag_ReadOnly = 1 << 16
 	};
 
-	enum TileOpacity {
-		// Image is full opaque and can be blitted fast
-		Opaque,
-		// Image has alpha and needs an alpha blit
-		Partial,
-		// Image is complately transparent
-		Transparent
-	};
 
 	/**
 	 * Provides opacity information about the image.
@@ -166,7 +158,7 @@ public:
 	 *
 	 * @return opacity information
 	 */
-	TileOpacity GetOpacity() const;
+	ImageOpacity GetImageOpacity() const;
 
 	/**
 	 * Provides opacity information about a tile on a tilemap.
@@ -177,7 +169,7 @@ public:
 	 *
 	 * @return opacity information
 	 */
-	TileOpacity GetTileOpacity(int row, int col) const;
+	ImageOpacity GetTileOpacity(int row, int col) const;
 
 	/**
 	 * Writes PNG converted bitmap to output stream.
@@ -538,13 +530,13 @@ public:
 	int bpp() const;
 	int pitch() const;
 
-protected:
-	TileOpacity CheckOpacity(Rect const& rect);
+	ImageOpacity ComputeImageOpacity(Rect rect) const;
 
+protected:
 	DynamicFormat format;
 
-	std::vector<std::vector<TileOpacity>> tile_opacity;
-	TileOpacity opacity = Partial;
+	std::vector<std::vector<ImageOpacity>> tile_opacity;
+	ImageOpacity image_opacity = ImageOpacity::Partial;
 	Color bg_color, sh_color;
 
 	friend void Text::Draw(Bitmap& dest, int x, int y, int color, FontRef font, std::string const& text, Text::Alignment align);
@@ -577,5 +569,21 @@ protected:
 	pixman_op_t GetOperator(pixman_image_t* mask = nullptr) const;
 	bool read_only = false;
 };
+
+inline ImageOpacity Bitmap::GetImageOpacity() const {
+	return image_opacity;
+}
+
+inline ImageOpacity Bitmap::GetTileOpacity(int row, int col) const {
+	return !tile_opacity.empty() ? tile_opacity[row][col] : ImageOpacity::Partial;
+}
+
+inline Color Bitmap::GetBackgroundColor() const {
+	return bg_color;
+}
+
+inline Color Bitmap::GetShadowColor() const {
+	return sh_color;
+}
 
 #endif