--- a/src/compiler.h
+++ b/src/compiler.h
@@ -21,10 +21,27 @@
 #ifdef __GNUC__
 #define EP_LIKELY(x) __builtin_expect(!!(x), 1)
 #define EP_UNLIKELY(x) __builtin_expect(!!(x), 0)
+
 #define EP_FALLTHROUGH __attribute__((fallthrough))
-#else
+
+#define EP_ALWAYS_INLINE __attribute__((always_inline)) inline
+
+#elif _MSC_VER
+
 #define EP_LIKELY(x) (x)
 #define EP_UNLIKELY(x) (x)
+
+#define EP_FALLTHROUGH do {} while (false)
+
+#define EP_ALWAYS_INLINE __forceinline
+
+#else
+
+#define EP_LIKELY(x) (x)
+#define EP_UNLIKELY(x) (x)
+
+#define EP_ALWAYS_INLINE inline
+
 #define EP_FALLTHROUGH do {} while (false)
 #endif