… bench/bitmap.cpp
… src/tilemap_layer.h