Retested #1417 

There was a performance regression introduced into tilemap when tone support was added. Basically we never actually use the `TileOpacity` value, resulting in us rendering every tile, even when they are fully transparent.

This PR fixes that and adds several other optimizations to TilemapLayer. 

The test case is a map with all transparent tiles and no background. Each optimization commit in this PR was measured and shown to have an improvement for this test.

For `ComputeImageOpacity()`

Master:
```
---------------------------------------------------------------------
Benchmark                              Time           CPU Iterations
---------------------------------------------------------------------
BM_ComputeImageOpacity            352849 ns     352851 ns       1969
BM_ComputeImageOpacityChipset     862583 ns     862591 ns        800
```

This PR:
```
---------------------------------------------------------------------
Benchmark                              Time           CPU Iterations
---------------------------------------------------------------------
BM_ComputeImageOpacity             70558 ns      70557 ns       9301
BM_ComputeImageOpacityChipset     116219 ns     116219 ns       5241
```

Master:
![tilemap_master](https://user-images.githubusercontent.com/1198466/71760982-44596180-2e93-11ea-97a9-51da82693c5e.png)
The calls to `operator()` are the div and mod lambdas in `TilemapLayer::Draw()`

This PR:
![tilemap](https://user-images.githubusercontent.com/1198466/71760936-e3318e00-2e92-11ea-9baa-e2e966d3918d.png)

Note in both profiles, the amount of time spent in `Game_picture::Update()` even though there are no pictures. This is fixed in #1989