@fmatthew5876 2020-01-04
Here is a profile of a map with grass and water tiles
![grassmap](https://user-images.githubusercontent.com/1198466/71761830-8df97a80-2e9a-11ea-9b9a-d3f65e1e9fe1.png)



Master:
![grass-old](https://user-images.githubusercontent.com/1198466/71761819-88039980-2e9a-11ea-8d7f-ce7a6c7e49cc.png)

This PR:
![grassnew](https://user-images.githubusercontent.com/1198466/71761854-92be2e80-2e9a-11ea-8984-22e964425e89.png)