@fmatthew5876 2020-01-05
I tested Witch's Heart which has a lot of crazy maps and a map with changing tone effects. Everything worked.