@Ghabry 2020-01-11
Another candidate for the "address some warnings" commit because you just added this line:

Clang reports:
```
easyrpg-player/src/game_actor.cpp:1234:4: warning: declaration does not declare anything [-Wmissing-declarations]
                        EP_FALLTHROUGH;
                        ^
easyrpg-player/src/compiler.h:24:24: note: expanded from macro 'EP_FALLTHROUGH'
#define EP_FALLTHROUGH __attribute__((fallthrough))
                       ^
```