@fmatthew5876 2020-01-15
Rebased again after mainloop1 PR.

I did a quick comparison of Frozen Triggers with vtune. Tilemap drawing is not a bottleneck in that game, but it does go from 0.5% to 0.2%.