@Ghabry 2020-01-15
@fmatthew5876 
This is easy to explain: The game uses a 1x1 map ;). I know this because I fixed the renderer for the less than 20x15 case :D